Update status bar volume for earbuds
I need to figure out a better solution for the status bar, and volume
control in general. It's hard to describe how terrible the current one
is. I've been working on another one which is not garbage, but the
status bar is obviously not high on my list of priorities. :-P